added mocha testing for getAllCustomers and separated customer queries into individual files
