+++ db-management/Customers/queries/deleteCustomer.js
@@ -1,0 +1,7 @@
+const db = require('../db-management/dbConfig');
+
+const deleteCustomer = (CustomerId) => {
+    db.serialize(() => {
+        db.run('DELETE FROM Customers WHERE CustomerId = ?', CustomerId);
+    });
+};
