+++ db-management/dbConfig.js
@@ -1,4 +1,4 @@
 const sqlite3 = require('sqlite3').verbose();
-const db = new sqlite3.Database('../db-management/bizeebird-old.db');
+const db = new sqlite3.Database('./db-management/bizeebird-old.db');
 
 module.exports = db;
