+++ package.json
@@ -4,7 +4,7 @@
   "description": "crm boarding app",
   "main": "main.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "mocha ./db-management/Customers/tests/getAllCustomersTEST.js",
     "start": "electron .",
     "start-debug": "DEBUG=TRUE electron .",
     "lint": "eslint \"**/*.js\"",
